hdf5 Optimizations!
* Adding large hdf5 chunk cache (20GB)
* Optimizing chunk size to be close to 1MB, but while retaining last dimension

+++ tests/time_test_hdf5.py
@@ -22,34 +22,38 @@ class Timer:
 
 
 def main():
-    n_arrays = 10
-
-    arrays = np.random.rand(n_arrays, 10, 500, 20, 600)
-    # arrays = (np.random.rand(n_arrays, 10, 500, 20, 600) > 0.5).astype(np.float)
-    # arrays = np.random.rand(n_arrays, 10, 5, 2, 6)
-
+    read_only = False
     data_dir = os.path.expanduser('~/output/tmp/hdf5-test')
-    os.makedirs(data_dir, exist_ok=True)
     file_pth = os.path.join(data_dir, 'data.h5')
-    if os.path.exists(file_pth):
-        os.remove(file_pth)
     key = 'train/what'
+    bs = 10
+    sts = 500 * 20
+    ns = 200
 
-    ## WRITE
-    hdf5_datastore = HDF5DataStore(file_pth)
-    recorder = Recorder(hdf5_datastore)
+    if not read_only:
+        n_arrays = 10
 
-    for i in range(n_arrays):
-        array = arrays[i]
-        with Timer() as st:
-            recorder.record(key, array)
-        if i == 0:
-            hdf5_datastore.enable_swmr()
-        print("Storing took %.2fs" % st.difftime)
-    recorder.close()
-    ## END WRITE
+        arrays = np.random.rand(n_arrays, bs, sts, ns)
 
-    print("File size after write is %d MiB" % (int(os.path.getsize(file_pth)) / 1024 / 1024))
+        os.makedirs(data_dir, exist_ok=True)
+        if os.path.exists(file_pth):
+            os.remove(file_pth)
+
+        ## WRITE
+        hdf5_datastore = HDF5DataStore(file_pth)
+        recorder = Recorder(hdf5_datastore)
+
+        for i in range(n_arrays):
+            array = arrays[i]
+            with Timer() as st:
+                recorder.record(key, array)
+            if i == 0:
+                hdf5_datastore.enable_swmr()
+            print("Storing took %.2fs" % st.difftime)
+        recorder.close()
+        ## END WRITE
+
+        print("File size after write is %d MiB" % (int(os.path.getsize(file_pth)) / 1024 / 1024))
 
     ## READ
     hdf5_datastore = HDF5DataStore(file_pth)
@@ -58,9 +62,18 @@ def main():
     with Timer() as rt:
         l = recorder.get_all(key)
     print("Reading took %.2fs" % rt.difftime)
+
+    for i in range(10):
+        b = np.random.randint(bs)
+        st = np.random.randint(sts)
+        with Timer() as rrt:
+            ll = np.array(l[:5, b, st, :])
+        print("Into sub-array took %.2f" % rrt.difftime)
+
     with Timer() as rrt:
         l = np.array(l)
     print("Into array took %.2f" % rrt.difftime)
+
     print("Mean is", np.mean(l), l.shape)
 
     recorder.close()
